Add native support for using multiple before and after images.
Before this change, the user had to create VRTs (https://gdal.org/drivers/raster/vrt.html) if they wanted to use multiple images as the before or after image of the example generation pipeline. This change allows users to directly specify multiple before and after images.

After this change, if a single building is covered by multiple pairs of before and after images, one example will be generated for each pair. How these examples are combined is currently up to the user. For instance, the user can keep all the examples and ensemble them, or rank the examples based on cloudiness, off-nadir angle, resolution, etc., and only keep the best one.

PiperOrigin-RevId: 486305400