Pass coordinates to dataflow workers using files instead of pipeline definition.
Previously, building coordinates were passed to workers using the beam.Create() function, which directly encodes the coordinates into the pipeline definition protobuf. This limits the number of coordinates that can be processed to less than 10 million, because the size of the pipeline definition protobuf is capped at 200MBs.

Using files, the number of coordinates can be as large as can fit into the memory of a single machine, which is at least 100 million.

This change also speeds up pipeline start up time.

PiperOrigin-RevId: 489551528