Adding string_label feature to tensorflow examples | Adding option to filter out labels from labeling task to create labeled dataset | Correcting string_to_numeric_labels default values
If labels_file is provided, the string_label feature is equal to the string_property value in labeled tensorflow examples.
If not, the string_label feature is equal to the value "no_label" in unlabeled tensorflow examples.

We exclude on-purpose labels that are not included in the string_to_numeric_flag to enable to user to explicitly include/exclude labels to create the labeled dataset.

Change "undamaged" default value to "no_damage" (consistency between create_labeling_task.py and create_labeled_dataset.py)