Add more parallelism to example generation pipeline.
New reprojection operations added in a recent CL are expensive and were taking a long time to run sequentially. This was especially true with a large number of images combined with a large number of buildings. This CL changes the example generation pipeline to execute the reprojections in parallel.

I also parallelized the window grouping operation to improve scalability. Window groups are now calculated in parallel for cells that are 0.1 degrees longitude by 0.1 degrees latitude, instead of sequentially for the whole image.

PiperOrigin-RevId: 642007026