Allows multiple labeled datasets to be created by assessment notebook.
Previously, only one labeled dataset was allowed, and newly generated datasets overwrote older ones.

With this change, labeled datasets are stored in a directory called "labeled_examples". Under this directory, each labeled dataset is stored in a subdirectory named as "{timestamp}_{train_fraction}_{test_fraction}_{minor_is_0}".

Models are now stored under the labeled_examples directory that they are trained on. This makes it easy to keep track of what data each model was trained on.

This change makes the notebook incompatible with output directories that were created by older versions of the notebook.

PiperOrigin-RevId: 703185458