Fix bug where pre and post images sometimes had a tilted black border around them.
The bug is triggered when the image is projected from the original CRS to the target UTM CRS. This projection will often introduce a little skew to the image, causing some pixels on the edge to be black.

This fix adds an extra margin of 10 pixels around each projected image. The black border is in most cases confined to this margin. The margin is then chopped off to obtain the final image, which will hopefully no longer have the border.

Note that in some extreme cases, this fix will not work. If the original image CRS is completely inappropriate (e.g. it's in the wrong UTM zone), then the projection will be extremely skewed and the black border will not be confined to the 10 pixel margin.

PiperOrigin-RevId: 747441416