Move post-processing steps in inference job to the Dataflow workers.
Previously, the post-processing steps ran on the machine that launches the inference job. When that machine is a free colab kernel, this setup is vulnerable to the colab kernel shutting down while the dataflow job is running and before the post-processing steps run.

By moving the post-processing steps to the Dataflow workers, this problem is avoided.

PiperOrigin-RevId: 756484558